% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{cm_5_8S}
\alias{cm_5_8S}
\title{Covariance Model for 5.8S rRNA}
\usage{
cm_5_8S()
}
\value{
(\code{character} string) the path to the CM file
}
\description{
This is a convenience function to return the path the RFAM 5.8S rRNA covariance model,
for use in \code{\link{cmsearch}} or \code{\link{cmalign}}.
The original file is from \url{https://rfam.xfam.org/family/RF00002/cm}.
}
\examples{
    # search sequences from a fasta file for Rfam RF00002 (5.8S rRNA)
    cm <- cm_5_8S()
    sampfasta <- sample_rRNA_fasta()
    cmsearch(cm = cm, seq = sampfasta, cpu = 1)
}
