% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferrnal.R
\name{cmalign}
\alias{cmalign}
\title{Align sequences to a covariance model}
\usage{
cmalign(cmfile, seq, glocal = TRUE, cpu)
}
\arguments{
\item{cmfile}{(\code{character} scalar) path to a covariance model file}

\item{seq}{(\code{character} scalar, names \code{character} vector, or
\code{\link[Biostrings]{XStringSet}}) sequences to align to the
covariance model. This may be given as a character path to a fasta
file, or the sequences as a character vector object of class
\code{\link[Biostrings]{DNAStringSet}} or
\code{\link[Biostrings]{RNAStringSet}}.  For \code{cmalign}, the
sequences should be known \emph{a priori} to represent the same region
as the CM; to find the region in longer sequences and align it, use
the \code{alignment} option of \code{\link{cmsearch}}.}

\item{glocal}{(\code{logical} scalar) If \code{TRUE}, align in glocal mode.
See \href{http://eddylab.org/infernal/}{Infernal} documentation for
more information.}

\item{cpu}{(\code{integer} scalar) The number of cpus to use.}
}
\value{
the aligned sequences, as returned by
    \code{\link{read_stockholm_msa}}.
}
\description{
This function calls \code{cmalign} from
\href{http://eddylab.org/infernal/}{Infernal}.  Infernal must be installed
and on the path.  Not all options are included.
}
\details{
One of the easiest places to obtain CMs is
\href{https://rfam.xfam.org/}{Rfam}.
}
\examples{
    # align a set of unaligned 5.8S rRNA sequences to the Rfam CM.
    cm5_8S <- system.file(
        file.path("extdata", "RF00002.cm"),
        package = "inferrnal"
    )
    unaln <- system.file(
        file.path("extdata", "samp_5_8S.fasta"),
        package = "inferrnal"
    )
    cmalign(cm5_8S, unaln, cpu = 1)
    # also works if the fasta file has already been loaded
    unaln <- Biostrings::readRNAStringSet(unaln)
    cmalign(cm5_8S, unaln, cpu = 1)
}
