% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferrnal.R
\name{cmbuild}
\alias{cmbuild}
\title{Construct covariance model(s) from structually annotated alignment(s)}
\usage{
cmbuild(
  msafile,
  cmfile_out,
  name = NULL,
  force = FALSE,
  summary_file = NULL,
  reannotated_msa = NULL,
  consensus_method = c("fast", "hand", "noss"),
  symfrac = NULL,
  rsearch = NULL,
  null = NULL,
  prior = NULL,
  weights = c("pb", "gsc", "given", "blosum"),
  wid = NULL,
  eff_num = c("ent", "none"),
  ere = NULL,
  eminseq = NULL,
  emaxseq = NULL,
  ehmmre = NULL,
  eset = NULL,
  p7ere = NULL,
  p7ml = FALSE,
  refine = NULL,
  local = FALSE,
  gibbs = FALSE,
  seed = NULL,
  cyk = FALSE,
  notrunc = FALSE,
  extra = NULL,
  verbose = FALSE
)
}
\arguments{
\item{msafile}{(\code{character} filename) filename of MSA file to read.}

\item{cmfile_out}{(\code{character} filename) filename to write CM to.}

\item{name}{(\code{character} string) name of the CM (option \code{-N} to cmbuild). The default uses the
name(s) given in the alignment file, or the name of the alignment file.}

\item{force}{(\code{logical} scalar) overwrite \code{cmfile_out} if it
exists}

\item{summary_file}{(\code{character} filename) filename to write summary
output to.}

\item{reannotated_msa}{(\code{character} filename) filename to write a
reannotated alignment(s) to (option \code{-O} to cmbuild)}

\item{consensus_method}{(one of \code{"fast"}, \code{"hand"}, or
\code{"noss"}) method to define consensus columns in alignment. (options
\code{--fast}, \code{--hand}, and \code{--noss} to cmbuild)}

\item{symfrac}{(\code{integer} scalar) residue fraction threshold necessary
to define a consensus column.}

\item{rsearch}{(\code{character} filename) RIBOSUM matrix to use in
parameterizing emission scores.}

\item{null}{(\code{character} filename) null model file.}

\item{prior}{(\code{character} filename) Dirichlet prior file.}

\item{weights}{(\code{character}; one of \code{"pb"}, \code{"gsc"},
\code{"none"}, \code{"given"}, or \code{"blosum"}) sequence weighting
method (options \code{--wpb}, \code{--wgsc},
\code{--wnone}, \code{--wgiven}, and \code{wblosum} to cmbuild).}

\item{wid}{(\code{numeric} scalar) percent identity for clustering when
\code{weights = "blosum"}.}

\item{eff_num}{(\code{character}; one of \code{"ent"} or \code{"none"})
entropy weighting strategy (options \code{--eent} or \code{--enone} to
cmbuild).}

\item{ere}{(\code{numeric} scalar) target mean match state relative entropy.}

\item{eminseq}{(\code{numeric} scalar) minimum allowed effective sequence
number.}

\item{emaxseq}{(\code{numeric} scalar) maximum allowed effective sequence
number.}

\item{ehmmre}{(\code{numeric} scalar) target HMM mean match state relative
entropy.}

\item{eset}{(\code{numeric} scalar) effective sequence number for entropy
weighting.}

\item{p7ere}{(\code{numeric} scalar) target mean match state relative entropy
for the filter p7 HMM.}

\item{p7ml}{(\code{logical} scalar) use a maximum liklihood p7 HMM built from
the CM.}

\item{refine}{(\code{character} filename) if given, the alignment is
iteratively refined, and the final alignment is written to this file.}

\item{local}{(\code{logical} scalar) use local alignment with \code{refine}
(option \code{-l} to cmbuild).}

\item{gibbs}{(\code{logical} scalar) use Gibbs sampling instead of
maximum likelihood with \code{refine}.}

\item{seed}{(\code{integer} scalar) pseudorandom number seed for Gibbs
sampling.}

\item{cyk}{(\code{logical} scalar) use the CYK alignment algorithm with
\code{refine}.}

\item{notrunc}{(\code{logical} scalar) turn off the truncated alignment
algorithm with \code{refine}.}

\item{extra}{(\code{character}) additional arguments to pass to cmbuild.}

\item{verbose}{(\code{logical} scalar) print cmbuild output to the console.}
}
\value{
\code{NULL}, invisibly
}
\description{
Calls the standalone program \code{cmbuild} from the
\href{http://eddylab.org/infernal/}{Infernal} package, which must be
installed.  For more information about options, see the
\href{http://eddylab.org/infernal/Userguide.pdf}{Infernal documentation}.
}
\examples{

# requires that LSUx is installed



}
