% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferrnal.R
\name{cmcalibrate}
\alias{cmcalibrate}
\title{Calibrate a covariance model for searching}
\usage{
cmcalibrate(
  cmfile,
  length = 1.6,
  forecast = FALSE,
  nforecast = NULL,
  memreq = FALSE,
  gtailn = 250,
  ltailn = 750,
  tailp = NULL,
  hfile = NULL,
  sfile = NULL,
  qqfile = NULL,
  ffile = NULL,
  xfile = NULL,
  seed = 181L,
  beta = 1e-15,
  nonbanded = FALSE,
  nonull3 = FALSE,
  random = FALSE,
  gc = NULL,
  cpu = NULL,
  mpi = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{cmfile}{(\code{character} file name) Filename of the CM to calibrate.
calibration data will be added to this file.}

\item{length}{(\code{numeric} scalar) Total length of random sequences to search,
in megabases. (parameter "\code{-L}" to \code{cmcalibrate})}

\item{forecast}{(\code{logical} flag) Predict the running time of
the calibration of cmfile (with provided options) on the current machine and
exit. The calibration is not performed.}

\item{nforecast}{(\code{integer} count) With \code{forecast}, specify the number of
processors that will be used for the calibration.}

\item{memreq}{(\code{logical} flag) Predict the amount of required memory for
calibrating cmfile (with provided options) on the current machine and exit.
The calibration is not performed.}

\item{gtailn}{(\code{numeric} scalar) Fit the exponential tail for glocal Inside
and glocal CYK to the \code{n} highest scores in the histogram tail, where \code{n} is
\code{gtailn} times \code{length} (the number of Mb searched).}

\item{ltailn}{(\code{numeric} scalar) Fit the exponential tail for local Inside
and local CYK to the \code{n} highest scores in the histogram tail, where \code{n} is
\code{ltailn} times \code{length} (the number of Mb searched).}

\item{tailp}{(\code{numeric} scalar) Ignore \code{gtailn} and \code{ltailn} options and fit
the \code{tailp} fraction tail of the histogram to an exponential tail, for all
search modes.}

\item{hfile}{(\code{character} file name) File to save the histograms fit.}

\item{sfile}{(\code{character} file name) File to save the survival plot
information.}

\item{qqfile}{(\code{character} file name) File to save the quantile-quantile plot
information.}

\item{ffile}{(\code{character} file name) File to save the statistics of
different exponential fits.}

\item{xfile}{(\code{character} file name) File to save a list of the scores in
each fit histogram.}

\item{seed}{(non-negative \code{integer} scalar) Random number seed. \code{0} seeds
the random number generator arbitrarily and non-reproducibly.}

\item{beta}{(non-negative \code{numeric} scalar) Beta tail loss probability used
for query-dependent banding (QDB).}

\item{nonbanded}{(\code{logical} flag) If \code{TRUE}, turn off QDB during E-value
calibration. This will slow down calibration.}

\item{nonull3}{(\code{logical} flag) If \code{TRUE}, turn off the null3 post hoc
additional null model. This is not recommended unless you plan on using the
same option to \code{cmsearch} and/or \code{cmscan}.}

\item{random}{(\code{logical} flag) If \code{TRUE}, use the background null model of
the CM to generate the random sequences, instead of the more realistic HMM.}

\item{gc}{(\code{character} file name) A file to use to determine the nucleotide
frequencies for the random sequences.}

\item{cpu}{(\code{integer} count) The number of parallel workers to use.}

\item{mpi}{(\code{logical} flag) If \code{TRUE}, run as an MPI parallel program.}

\item{quiet}{(\code{logical} flag) If \code{TRUE}, suppress standard output of
\code{cmcalibrate}.}
}
\value{
NULL, invisibly
}
\description{
Calibrate a covariance model for searching
}
