% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferrnal.R
\name{cmsearch}
\alias{cmsearch}
\title{Search for Covariance Models (CM) in a set of sequences.}
\usage{
cmsearch(cm, seq, glocal = TRUE, alignment = NULL, cpu = NULL)
}
\arguments{
\item{cm}{(character of length 1) Path to the covariance model (.cm) file.
The covariance model must include calibration data from running
"\code{cmcalibrate}".}

\item{seq}{(filename, character vector,
\code{\link[Biostrings]{XStringSet}}, or
\code{\link[ShortRead]{ShortRead}}) Sequences to search with the CM.
If a filename, the file can be of any type supported by Infernal.}

\item{glocal}{(logical of length 1) Whether to run the search in glocal mode
(global with respect to the model, local with respect to the sequence).
When \code{TRUE}, the search is faster, but will fail to find matches with
only partially homologous structure.}

\item{alignment}{(filename) A file to save the aligned hits to.  If given,
the alignment is saved in Stockholm format with annotations for secondary
structure, posterior probablility, etc.}

\item{cpu}{(integer of length 1) The number of threads to use in the search.}
}
\value{
a \code{\link[tibble]{tibble}} with columns:
    \itemize{
        \item{target_name}{ (character) the name of the target sequence}
        \item{taget_accession}{(character) the target's accession number}
        \item{query_name}{(character) the name of the query CM}
        \item{query_accession}{(character) the query CM's accession number}
        \item{mdl}{(character) the model type ("cm" or "hmm")}
        \item{mdl_from}{(integer) the start location of the hit in the model}
        \item{mdl_to}{(integer) the end location of the hit in the model}
        \item{seq_from}{(integer) the start location of the hit in the
            sequence}
        \item{seq_to}{(integer) the end location of the hit in the sequence}
        \item{strand}{(character) the strand the hit was found on ("+" or
            "-")}
        \item{trunc)}{(character) whether the hit is truncated, and where
            ("no", "5'", "3'", "5'&3'", or "-" for hmm hits).}
        \item{pass}{(integer) which algorithm pass the hit was found on.}
        \item{gc}{(numeric) GC content of the hit}
        \item{bias}{(numeric) biased composition correction.  See the
            Infernal documentation.}
        \item{score}{(numeric) bit-score of the hit, including the biased
            composition correction.}
        \item{E_value}{(numeric) Expectation value for the hit.}}
        \item{inc}{(character) "!" if the sequence meets the inclusion
            threshold, "?" if it only meets the reporting threshold.}
}
\description{
This function calls "\code{cmsearch}" from
\href{http://eddylab.org/infernal/}{Infernal}.  Infernal must be installed.
Many parameters are not included (yet!), and the function is focused on
retrieving the hits table and, optionally, producing an alignment.
}
\examples{
    # search sequences from a fasta file for Rfam RF00002 (5.8S rRNA)
    cm <- cm_5_8S()
    sampfasta <- sample_rRNA_fasta()
    cmsearch(cm = cm, seq = sampfasta, cpu = 1)
    # also works if the fasta file has already been loaded
    samp <- Biostrings::readDNAStringSet(sampfasta)
    cmsearch(cm = cm, seq = samp, cpu = 1)
}
