% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rfam_cm.R
\name{download_rfam_cm}
\alias{download_rfam_cm}
\title{Download a covariance model from RFAM}
\usage{
download_rfam_cm(cm, filename = paste0(cm, ".cm"), ...)
}
\arguments{
\item{cm}{(\code{character} string) the name of the CM to download. Typically
this is in the format \code{"RFxxxxx"} where the \code{"x"}'s represent
decimal digits.}

\item{filename}{(\code{character} string) file to save the CM in.  The
default is to save the CM in the current working directory in a file
named \code{"RFxxxxx.cm"}.}

\item{...}{additional arguments passed on to
\code{\link[utils]{download.file}}.}
}
\value{
The name of the downloaded file, invisibly.
}
\description{
Download a covariance model from RFAM
}
