% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stockholm_msa.R
\name{read_stockholm_msa}
\alias{read_stockholm_msa}
\title{Parse a Multiple Alignment in Stockholm Format}
\usage{
read_stockholm_msa(stockholm, type = c("RNA", "DNA", "AA"))
}
\arguments{
\item{stockholm}{(\code{character} scalar) Path to a file to parse}

\item{type}{(\code{character} scalar) Type of alignment; "RNA", "DNA", or "AA".}
}
\value{
a \code{\link[=StockholmMultipleAlignment-class]{StockholmMultipleAlignment}}
}
\description{
Parses Stockholm-format multiple alignment files, including "\verb{#=GC}"
lines.  Other annotations are ignored.
}
\examples{
    msafile <- sample_rRNA_stk()
    msa <- read_stockholm_msa(msafile)
    msa
}
