% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stockholm_msa.R
\name{writeStockholmMultipleAlignment}
\alias{writeStockholmMultipleAlignment}
\title{Write an annotated alignment in Stockholm format}
\usage{
writeStockholmMultipleAlignment(x, connection)
}
\arguments{
\item{x}{(\code{\link[=StockholmMultipleAlignment-class]{StockholmMultipleAlignment}})
the object to write to the file}

\item{connection}{(\code{\link[base:connections]{connection}} object or file name)
the file to write the alignment to}
}
\value{
the connection, invisibly
}
\description{
Write an annotated alignment in Stockholm format
}
