% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stockholm_msa.R
\name{read_stockholm_msa}
\alias{read_stockholm_msa}
\title{Parse a Multiple Alignment in Stockholm Format}
\usage{
read_stockholm_msa(stockholm, dna = FALSE)
}
\arguments{
\item{stockholm}{(\code{character} scalar) Path to a file to parse}

\item{dna}{(\code{logical} scalar) Parse the input as DNA instead of RNA.}
}
\value{
a \code{list}, with elements:
    \describe{
        \item{\code{alignment}
            (\code{\link[Biostrings]{MultipleAlignment}})}{
            The alignment itself}
        \item{\code{GF} (\code{character})}{Annotations which apply to the
            entire file.}
        \item{\code{GS} (\code{\link[S4Vectors]{DataFrame}} containing one
            \code{\link[Biostrings]{BStringSet}} for each annotation)}{
            Annotations which apply to each sequence.}
        \item{\code{GC} (\code{list} of \code{\link[Biostrings]{BString}})}{
            Annotations which apply to each column of the alignment;
            in particular, "\code{SS_cons}" is the consensus secondary
            structure, and "\code{RF}" is the reference line.}
        \item{\code{GR} (\code{\link[S4Vectors]{DataFrame}} containing one
            \code{\link[Biostrings]{BStringSet}} for each annotation)}{%
            Annotations which apply to each residue (i.e., each column and
            sequence) in the alignment.}
    }
}
\description{
Parses Stockholm-format multiple alignment files, including "\code{#=GC}"
lines.  Other annotations are ignored.
}
\examples{
    msafile <- sample_rRNA_stk()
    msa <- read_stockholm_msa(msafile)
    msa$alignment
    # consensus secondary structure
    msa$SS_cons
    # reference sequence
    msa$RF
}
