% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferrnal.R
\name{read_stockholm_msa}
\alias{read_stockholm_msa}
\title{Parse a Multiple Alignment in Stockholm Format}
\usage{
read_stockholm_msa(stockholm)
}
\arguments{
\item{stockholm}{(\code{character} scalar) Path to a file to parse}
}
\value{
a \code{list}.  The sequences themselves are element "\code{alignment}",
    and \code{#=GC} annotations are given as \code{character} strings; e.g.,
    the "\code{#=GC SS_cons}" annotation is stored as an element named
    "\code{SS_cons}".
}
\description{
Parses Stockholm-format multiple alignment files, including "\code{#=GC}"
lines.  Other annotations are ignored.
}
\examples{
    msafile <- system.file(file.path("extdata", "sample.stk"), package = "inferrnal")
    msa <- read_stockholm_msa(msafile)
    msa$alignment
    # consensus secondary structure
    msa$SS_cons
    # reference sequence
    msa$RF
}
