% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockholm_msa.R
\docType{class}
\name{StockholmMultipleAlignment-class}
\alias{StockholmMultipleAlignment-class}
\alias{StockholmDNAMultipleAlignment}
\alias{StockholmRNAMultipleAlignment}
\alias{StockholmAAMultipleAlignment}
\title{StockholmMultipleAlignment objects}
\usage{
StockholmDNAMultipleAlignment(
  x = character(),
  start = NA,
  end = NA,
  width = NA,
  use.names = TRUE,
  rowmask = NULL,
  colmask = NULL,
  GF = character(),
  GS = list(),
  GR = list(),
  GC = character()
)

StockholmRNAMultipleAlignment(
  x = character(),
  start = NA,
  end = NA,
  width = NA,
  use.names = TRUE,
  rowmask = NULL,
  colmask = NULL,
  GF = character(),
  GS = list(),
  GR = list(),
  GC = character()
)

StockholmAAMultipleAlignment(
  x = character(),
  start = NA,
  end = NA,
  width = NA,
  use.names = TRUE,
  rowmask = NULL,
  colmask = NULL,
  GF = character(),
  GS = list(),
  GR = list(),
  GC = character()
)
}
\arguments{
\item{x}{(\code{character} vector, aligned
\code{\link[Biostrings:XStringSet-class]{XStringSet}} or
\code{\link[Biostrings:MultipleAlignment-class]{MultipleAlignment}} of an appropriate
type) multiple sequence alignment without Stockholm extensions}

\item{start, end, width, use.names, rowmask, colmask}{passed to the appropriate
\code{\link[Biostrings:MultipleAlignment-class]{MultipleAlignment}} constructor,
unless "x" is already a
\code{\link[Biostrings:MultipleAlignment-class]{MultipleAlignment}}}

\item{GF}{(named \code{character} vector or
\code{\link[Biostrings:XStringSet-class]{BStringSet}}) Free-text annotations
which belong to the alignment file as a whole.  The name of each element is
a tag identifying the type of data. (See Details).}

\item{GS}{(named \code{list} of named \code{character} vectors or
\code{\link[Biostrings:XStringSetList-class]{BStringSetList}}) Free-text
annotations which belong to the individual sequences in the alignment. The
names of the outer \code{list} or
\code{\link[Biostrings:XStringSetList-class]{BStringSetList}} are tags
identifying the type of data for each element. (See Details). Names of inner
\code{character} vectors or \code{\link[Biostrings:XStringSet-class]{BStringSet}}s match
the names of sequences in the alignment, but there is no requirement that
every sequence must be annotated for every tag.}

\item{GR}{(named \code{list} of named \code{character} vectors or
\code{\link[Biostrings:XStringSetList-class]{BStringSetList}}) Annotations
for individual residues in the alignment. The names of the outer \code{list} or
\code{\link[Biostrings:XStringSetList-class]{BStringSetList}} are tags
identifying the type of data for each element. (See Details). Names of inner
\code{character} vectors or \code{\link[Biostrings:XStringSet-class]{BStringSet}}s match
the names of sequences in the alignment, but there is no requirement that
every sequence must be annotated for every tag. Unlike GS tags, the width of
all elements must be the same, and must match the width of the alignment.}

\item{GC}{(named \code{character} vector or
\code{\link[Biostrings:XStringSet-class]{BStringSet}}) Annotations
which belong to each column of the alignment as a whole. The name of each
element is a tag identifying the type of data. (See Details).  Unlike GF
tags, the width of all elements must be the same, and must match the width
of the alignment.}
}
\value{
a new \code{StockholmMultipleAlignment} object
}
\description{
The \code{StockholmMultipleAlignment} class contains a multiple sequence
alignment along with its annotations, as defined for the Stockholm file
format.
}
\details{
Although the \code{StockholmMultipleAlignment} class is agnostic about the
specific tags used, the following tags are the most likely to be recognized
by Infernal or other software which reads or writes Stockholm files:

\tabular{lll}{
Type \tab Tag \tab Description \cr
GF \tab ID \tab IDentifier \cr
GF \tab AC \tab ACcession \cr
GF \tab DE \tab DEscription \cr
GF \tab AU \tab AUthor \cr
GF \tab GA \tab GAthering threshold \cr
GF \tab NC \tab Noise Cutoff \cr
GF \tab TC \tab Trusted Cutoff \cr
GS \tab WT \tab WeighT \cr
GS \tab AC \tab ACcession number \cr
GS \tab DE \tab DEscription \cr
GS \tab DR \tab Database Reference \cr
GS \tab OS \tab OrganiSm (species) \cr
GS \tab OC \tab Organism Classification (clade, etc.) \cr
GS \tab LO \tab Look (Color, etc.) \cr
GR \tab SS \tab Secondary Structure \cr
GR \tab SA \tab Surface Accessibility \cr
GR \tab TM \tab TransMembrane \cr
GR \tab PP \tab Posterior Probability \cr
GR \tab LI \tab LIgand binding \cr
GR \tab AS \tab Active Site \cr
GR \tab pAS \tab AS - Pfam predicted \cr
GR \tab sAS \tab AS - from SwissProt \cr
GR \tab IN \tab INtron (in or after) \cr
GC \tab RF \tab ReFerence \cr
GC \tab SS_cons \tab Secondary Structure consensus \cr
GC \tab SA_cons \tab Surface Accessibility consensus \cr
GC \tab TM_cons \tab TransMembrane consensus \cr
GC \tab PP_cons \tab Posterior Probability consensus \cr
GC \tab LI_cons \tab LIgand binding consensus \cr
GC \tab AS_cons \tab Active Site consensus \cr
GC \tab pAS_cons \tab AS - Pfam predicted consensus \cr
GC \tab sAS_cons \tab AS - from SwissProt consensus \cr
GC \tab IN_cons \tab INtron (in or after) consensus \cr
}
}
\section{Slots}{

\describe{
\item{\code{GF}}{\code{\link[Biostrings:XStringSet-class]{BStringSet}}. Free-text annotations
which belong to the alignment file as a whole.  The name of each element is
a tag identifying the type of data. (See Details).}

\item{\code{GS}}{\code{\link[Biostrings:XStringSetList-class]{BStringSetList}}. Free-text
annotations which belong to the individual sequences in the alignment. The
name of each \code{\link[Biostrings:XStringSet-class]{BStringSet}} is a tag
identifying the type of data. (See Details). Names of individual
\code{\link[Biostrings:XString-class]{BString}} elements match the names of sequences
in the alignment, but there is no requirement that every sequence must be
annotated for every tag.}

\item{\code{GR}}{\code{\link[Biostrings:XStringSetList-class]{BStringSetList}}. Annotations
for individual residues in the alignment. The name of each
\code{\link[Biostrings:XStringSet-class]{BStringSet}} is a tag identifying the type
of data. (See Details). Names of individual
\code{\link[Biostrings:XString-class]{BString}} elements match the names of sequences
in the alignment, but there is no requirement that every sequence must be
annotated for every tag. Unlike GS tags, the width of all elements must be
the same, and must match the width of the alignment.}

\item{\code{GC}}{\code{\link[Biostrings:XStringSet-class]{BStringSet}}. Annotations
which belong to each column of the alignment as a whole. The name of each
element is a tag identifying the type of data. (See Details).  Unlike GF
tags, the width of all elements must be the same, and must match the width
of the alignment.}
}}

\examples{
# Typically a StockholmMultipleAlignment object is read from a file created
# by other software, but it can also be created manually.
# This example reproduces the example file given in the Stockholm format
# definition.
samp <- StockholmAAMultipleAlignment(
    x = c(
        "O83071/192-246" = "MTCRAQLIAVPRASSLAE..AIACAQKM....RVSRVPVYERS",
        "O83071/259-312" = "MQHVSAPVFVFECTRLAY..VQHKLRAH....SRAVAIVLDEY",
        "O31698/18-71"   = "MIEADKVAHVQVGNNLEH..ALLVLTKT....GYTAIPVLDPS",
        "O31698/88-139"  = "EVMLTDIPRLHINDPIMK..GFGMVINN......GFVCVENDE",
        "O31699/88-139"  = "EVMLTDIPRLHINDPIMK..GFGMVINN......GFVCVENDE"
    ),
    GF = c(
        ID = "CBS",
        AC = "PF00571",
        AU = "Bateman A",
        CC = paste("CBS domains are small intracellular modules mostly",
                   "found in 2 or four copies within a protein."),
        SQ = "67"
    ),
    GS = list(
        # ACcession number
        AC = c(
            "O31698/18-71" = "O31698",
            "O83071/192-246" = "O83071",
            "O83071/259-312" = "O83071",
            "O31698/88-139" = "O31698"
        ),
        # OrganiSm
        OS = c("O31698/88-139" = "Bacillus subtilis")
    ),
    GR = list(
        # Surface Accessibility
        SA = c(
            "O83071/192-246" = "999887756453524252..55152525....36463774777"
        ),
        # Secondary Structure
        SS = c(
            "O83071/259-312" = "CCCCCHHHHHHHHHHHHH..EEEEEEEE....EEEEEEEEEEE",
            "O31698/18-71"   = "CCCHHHHHHHHHHHHHHH..EEEEEEEE....EEEEEEEEHHH",
            "O31698/88-139"  = "CCCCCCCHHHHHHHHHHH..HEEEEEEE....EEEEEEEEEEH"
        ),
        # Active Site
        AS = c(
            "O31699/88-139"  = "________________*__________________________"
        ),
        # INtron
        IN = c(
            "O31699/88-139"  = "____________1______________2__________0____"
        )
    ),
    GC = c(
        # Secondary Structure consensus
        SS_cons = "CCCCCHHHHHHHHHHHHH..EEEEEEEE....EEEEEEEEEEH"
    )
 )
 samp
}
